requirements.txt
streamlit==1.36.0
torch==2.1.2
transformers==4.37.2
Pillow==10.2.0
numpy==1.26.4
